% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv.R
\name{create_cv}
\alias{create_cv}
\title{Create Cross Validate ID's}
\usage{
create_cv(
  space.id,
  time.id,
  num.folds = 10,
  type = "spatial",
  coords = NULL,
  buffer.size = NULL
)
}
\arguments{
\item{space.id}{Spatial location ID vector (n)}

\item{time.id}{Temporal location ID vector (n)}

\item{num.folds}{Number of folds used in the cross validation process (default = 10)}

\item{type}{Type of cross validation to be performed. Options are "spatial", "ordinary", or "spatial_clustered". (default = "spatial")}

\item{coords}{Matrix of x y coordinates, with colnames(coords) == c("x", "y"), (n, 2)}

\item{buffer.size}{Radius of buffer size, if type = "spatial_buffered" (default = 0)}
}
\value{
A named list containing a vector of cross validation ID's, number of folds, and cross validation type.
}
\description{
This function creates a vector of cross validation ID's for a given number of folds and type of cross validation.
}
\examples{
# grm_cv()


}
