% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv.R
\name{create_cv_original}
\alias{create_cv_original}
\title{Create Cross Validate ID's}
\usage{
create_cv_original(
  time.id,
  space.id,
  spacetime.id,
  num.folds = 10,
  type = "spatial",
  coords = NULL,
  buffer.size = NULL
)
}
\arguments{
\item{time.id}{Temporal location ID vector (n)}

\item{space.id}{Spatial location ID vector (n)}

\item{spacetime.id}{ID vector of time points where spatial trends vary (n)}

\item{num.folds}{Number of folds used in the cross validation process (default = 10)}

\item{type}{Type of cross validation to be performed. Options are "spatial", "ordinary", "spatial_clustered", or "spatial_buffered". (default = "spatial")}

\item{coords}{Matrix of x y coordinates, with colnames(coords) == c("x", "y"), (n, 2)}

\item{buffer.size}{Radius of buffer size, if type = "spatial_buffered" (default = 0)}
}
\value{
A named list containing a vector of cross validation ID's, a matrix of which observations to drop for each fold if the cv type is "spatial_buffered", and inputted objects.
}
\description{
This function creates a vector of cross validation ID's for a given number of folds and type of cross validation.
}
