% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_spatial.R
\name{ensemble_spatial}
\alias{ensemble_spatial}
\title{Fit the Spatial Ensemble model}
\usage{
ensemble_spatial(
  grm.fit.cv.1,
  grm.fit.cv.2,
  date.Y.1,
  date.Y.2,
  coords.Y.1,
  space.id.Y.1,
  n.iter = 25000,
  burn = 5000,
  thin = 4,
  tau.a = 0.001,
  tau.b = 0.001,
  theta.tune = 0.2,
  theta.a = 5,
  theta.b = 0.05
)
}
\arguments{
\item{grm.fit.cv.1}{Output from grm_cv function for first primary variable}

\item{grm.fit.cv.2}{Output from grm_cv function for second primary variable}

\item{date.Y.1}{Date (or other unique day id) for first primary variable}

\item{date.Y.2}{Date (or other unique day id) for second primary variable}

\item{coords.Y.1}{Matrix of x y coordinates for first primary variable, with colnames(coords) == c("x", "y"), (n, 2)}

\item{space.id.Y.1}{Spatial location ID vector for first primary variable (n)}

\item{n.iter}{Number of iterations used in predictions. Must be <= than post-thined and burned iterations from grm.fit}

\item{burn}{Number of pre-covergence simulations}

\item{thin}{Save every thin'th simulation}

\item{tau.a}{First tau prior hyperparameter}

\item{tau.b}{Second tau prior hyperparameter}

\item{theta.tune}{Theta Metropolis-Hastings proposal tuning parameter}

\item{theta.a}{First theta prior hyperparameter}

\item{theta.b}{Second theta prior hyperparameter}
}
\value{
A data frame containing cross validation predictions
}
\description{
This function fits the spatial ensemble, using output from the grm_cv function applied to multiple data sets
}
\examples{
# ensemble_spatial()


}
