% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_fcn.R
\name{first_fcn}
\alias{first_fcn}
\title{Print a test string}
\usage{
first_fcn(test)
}
\arguments{
\item{test}{A character vector to print}
}
\value{
A string
}
\description{
Print a test string
}
\examples{
test <- "this is a test"
first_fcn(test)
}
