% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_fill.R
\name{gap_fill}
\alias{gap_fill}
\title{Make predictions for locations with two observations}
\usage{
gap_fill(Y.pred.1, Y.pred.2, Y.sd.1, Y.sd.2, space.id, weights)
}
\arguments{
\item{Y.pred.1}{Predictions made with grm_pred and first primary independent variable}

\item{Y.pred.2}{Predictions made with grm_pred and second primary independent variable}

\item{Y.sd.1}{Prediction standard deviation output from grm_pred and first primary independent variable}

\item{Y.sd.2}{Prediction standard deviation output from grm_pred and second primary independent variable}

\item{space.id}{Spatial location ID vector (n)}

\item{weights}{Ensemble weights output from weight_pred}
}
\value{
A list including ensemble predictions and standard deviations
}
\description{
This functions uses the weights and predictions from the previous separate grm's to make ensemble predictions
}
\examples{
# gap_fill()


}
