% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv.R
\name{get_spat_cv_id}
\alias{get_spat_cv_id}
\title{Get Spatial Cross Validate ID's For Regular Spatial Cross Validation}
\usage{
get_spat_cv_id(space_id_cv, num.folds)
}
\arguments{
\item{space_id_cv}{Spatial ID's for cross validation}

\item{num.folds}{Number of folds used in the cross validation process (default = 10)}
}
\value{
A vector of cross validation ID's
}
\description{
This function creates a vector of spatial cross validation ID's, used in both spatial cv and spatial_buffered cv, in main create_cv() function.
}
