% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm.R
\name{grm}
\alias{grm}
\title{Fit the Geostatistical Regression Model (GRM)}
\usage{
grm(
  Y,
  X,
  L,
  M,
  dist.space.mat,
  space.is,
  time.id,
  spacetime.id,
  t1.a = 0.5,
  t1.b = 0.005,
  theta.tune = 0.2,
  theta.a = 5,
  theta.b = 0.05,
  sigma.a = 0.001,
  sigma.b = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Dependent variable vector (n).}

\item{X}{Primary inependent variable vector (n).}

\item{L}{Spatial covariate matrix (n, p1)}

\item{M}{Spatial covariate matrix (n, p2)}
}
\value{
A string
}
\description{
This function fits Bayesian Hierarchical Model (BHM) in the form of \eqn{Y \sim \beta X + \gamma L + \delta M}
}
\examples{
test <- "this is a test"
first_fcn(test)
}
