% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_pred.R
\name{grm_pred}
\alias{grm_pred}
\title{Fit the Geostatistical Regression Model (GRM) With Cross Validation}
\usage{
grm_pred(
  grm.fit,
  X.pred,
  L.pred,
  M.pred,
  locations.Y,
  locations.pred,
  space.id,
  time.id,
  spacetime.id,
  include.additive.annual.resid = T,
  include.multiplicative.annual.resid = T,
  n.iter = 500,
  verbose = TRUE
)
}
\arguments{
\item{grm.fit}{Fit object created with grm()}

\item{X.pred}{Standardized primary independent variable vector for all predictions (n_pred)}

\item{L.pred}{Standardized spatial covariate matrix for all predictions (n_pred, p1)}

\item{M.pred}{Standardized spatio-temporal covariate matrix for all predictions (n_pred, p2)}

\item{locations.Y}{Matrix of primary independent variable x, y coodinates (n_obs, 2)}

\item{locations.pred}{Matrix of prediction x, y coodinates (n_pred, 2)}

\item{space.id}{Spatial location ID vector (n)}

\item{time.id}{Temporal location ID vector (n)}

\item{spacetime.id}{ID vector of time points where spatial trends vary (n)}

\item{include.additive.annual.resid}{Include additive spatial (annual) residual bias not explained by other inputes}

\item{include.multiplicative.annual.resid}{Include multiplicative spatial (weekly) residual bias not explained by other inputes}

\item{n.iter}{Number of iterations used in predictions. Must be <= than post-thined and burned iterations from grm.fit}

\item{verbose}{Print MCMC output}
}
\value{
A data frame containing cross validation predictions
}
\description{
This function fits Bayesian Hierarchical Model (BHM) in the form of Y ~ beta X + gamma L + delta M with cross-validation
}
\examples{
# grm_pred()


}
