% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maia_pm25}
\alias{maia_pm25}
\title{PM2.5 Monitor and Corresponding Satellite Aerial Optical Depth Data For Atlanta Metro Area, 2003-2005}
\format{
\subsection{\code{maia_pm25}}{

A data frame with 6531 rows and 20 columns:
\describe{
\item{time_id}{Week identifier}
\item{space_id}{Location identifier}
\item{spacetime_id}{Year identifier}
\item{pm25}{PM2.5 reading}
\item{aod}{Aerial optical depth value}
\item{elevation}{Elevation spatial covariate (US Geological Survey)}
\item{forestcover}{Percentage of forest cover spatial covariate (2001 National Land Cover database)}
\item{hwy_length}{Sum of major roadway length spatial covariate (2001 National Land Cover database)}
\item{lim_hwy_length}{Sum of major roadway length spatial covariate (2001 National Land Cover database)}
\item{lim_hwy_length}{Sum of local roadway lengths spatial covariate (2001 National Land Cover database)}
\item{point_emi_any}{Indicator of PM2.5 primary emission point source spatial covariate (2002 USEPA National Emissions Inventory)}
\item{tmp}{Temperature spatio-temporal covariate (North American Land Data Assimilation Systems)}
\item{wind}{Wind spatio-temporal covariate (North American Land Data Assimilation Systems)}
\item{cmaq}{CMT value spatio-temporal covariate (USEPA Models-3/Community Multiscale Air Quality)}
\item{date}{Date of measurement}
\item{x}{Location x-coordinate}
\item{y}{Location y-coordinate}
...
}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31465992/}
}
\usage{
maia_pm25
}
\description{
Daily PM2.5 readings, AOD, and covariates of interest in Atlanta Metro Area from 2003-2005.
Each row of the data is a unique location/day.
Data included only if AOD output is available for that day and monitoring location.
(ie. if the monitor is within a 12km x 12km grid square for which AOD data is available).
}
\keyword{datasets}
