% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred.R
\name{weight_pred}
\alias{weight_pred}
\title{Generate weights from ensemble spatial fit}
\usage{
weight_pred(q, theta, tau2, locations.Y, locations.pred, verbose = TRUE)
}
\arguments{
\item{q}{Matrix of samples output from ensemble_spatial function}

\item{theta}{Vector of theta samples output from ensemble_spatial function}

\item{tau2}{Vector of tau2 samples output from ensemble_spatial function}

\item{locations.Y}{Matrix of primary independent variable x, y coodinates (n_obs, 2)}

\item{locations.pred}{Matrix of prediction x, y coodinates (n_pred, 2)}

\item{verbose}{Print MCMC output}
}
\value{
A matrix containing weights
}
\description{
This function generates weights using the output from the ensemble_spatial function
}
\examples{
# grm_pred()


}
