% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_spatial.R
\name{ensemble_spatial}
\alias{ensemble_spatial}
\title{Fit the Spatial Ensemble model}
\usage{
ensemble_spatial(
  d1,
  d2,
  dist.space.mat,
  space.id,
  n.iter = 25000,
  burn = 5000,
  thin = 4,
  tau.a = 0.001,
  tau.b = 0.001,
  theta.tune = 0.2,
  theta.a = 5,
  theta.b = 0.05
)
}
\arguments{
\item{d1}{Vector of densities from first dataset grm_cv() output}

\item{d2}{Vector of densities from second dataset grm_cv() output}

\item{dist.space.mat}{Spatial distance matrix}

\item{space.id}{Spatial location ID vector (n)}

\item{n.iter}{Number of iterations used in predictions. Must be <= than post-thined and burned iterations from grm.fit}

\item{burn}{Number of pre-covergence simulations}

\item{thin}{Save every thin'th simulation}

\item{tau.a}{First tau prior hyperparameter}

\item{tau.b}{Second tau prior hyperparameter}

\item{theta.tune}{Theta Metropolis-Hastings proposal tuning parameter}

\item{theta.a}{First theta prior hyperparameter}

\item{theta.b}{Second theta prior hyperparameter}
}
\value{
A data frame containing cross validation predictions
}
\description{
This function fits the spatial ensemble, using output from the grm_cv function applied to multiple data sets
}
\examples{
# ensemble_spatial()


}
