% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_fill.R
\name{gap_fill}
\alias{gap_fill}
\title{Make predictions for locations with two observations}
\usage{
gap_fill(grm.pred.1, grm.pred.2, date.pred.1, date.pred.2, space.id, weights)
}
\arguments{
\item{grm.pred.1}{Output from grm_pred() function for first primary variable}

\item{grm.pred.2}{Output from grm_pred() function for second primary variable}

\item{date.pred.1}{Date (or other unique day identifier) from first variable prediction dataset}

\item{date.pred.2}{Date (or other unique day identifier) from second variable prediction dataset}

\item{space.id}{Spatial location ID vector (n)}

\item{weights}{Ensemble weights output from weight_pred}
}
\value{
A list including ensemble predictions and standard deviations
}
\description{
This functions uses the weights and predictions from the previous separate grm's to make ensemble predictions
}
\examples{
# gap_fill()


}
