% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred.R
\name{weight_pred}
\alias{weight_pred}
\title{Generate weights from ensemble spatial fit}
\usage{
weight_pred(
  ensemble.fit,
  coords.Y.1,
  space.id.Y.1,
  coords.pred.1,
  space.id.pred.1,
  verbose = TRUE
)
}
\arguments{
\item{ensemble.fit}{Fit from ensemble_spatial() function}

\item{coords.Y.1}{Coordinate matrix for first primary variable}

\item{space.id.Y.1}{Space id for first primary variable}

\item{coords.pred.1}{Coordinate matrix for first variable predictions}

\item{space.id.pred.1}{Space id for first primary variable predictions}

\item{verbose}{Print MCMC output}
}
\value{
A matrix containing weights
}
\description{
This function generates weights using the output from the ensemble_spatial function
}
\examples{
# weight_preds()


}
